Correção de erros de excluir e editar veiculo;

+++ src/main/java/br/com/unialfa/ecomerce/carro/domain/carro.java
@@ -5,14 +5,13 @@ import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import java.io.Serializable;
-import java.math.BigDecimal;
 
 @Entity
 public class carro implements Serializable {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private int id;
+    private long id;
     private String marca;
     private String modelo;
     private String cor;
@@ -23,11 +22,11 @@ public class carro implements Serializable {
     public carro() {
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
